+++ source/Examples/Resources/CMFallbackStatusPoint/CMFallbackStatusPoint_Absent.ps1
@@ -1,0 +1,18 @@
+<#
+    .SYNOPSIS
+        A DSC configuration script to remove a fallback status point from Configuration Manager.
+#>
+Configuration Example
+{
+    Import-DscResource -ModuleName ConfigMgrCBDsc
+
+    Node localhost
+    {
+        CMFallbackStatusPoint ExampleSettings
+        {
+            SiteCode       = 'Lab'
+            SiteServerName = 'FSP01.contoso.com'
+            Ensure         = 'Absent'
+        }
+    }
+}